+++ api/src/main/java/org/allaymc/api/entity/interfaces/EntityArmorStand.java
@@ -1,7 +1,14 @@
 package org.allaymc.api.entity.interfaces;
 
-import org.allaymc.api.entity.Entity;
+import org.allaymc.api.entity.component.EntityArmorStandBaseComponent;
+import org.allaymc.api.entity.component.EntityContainerHolderComponent;
+import org.allaymc.api.entity.component.EntityPhysicsComponent;
 
-public interface EntityArmorStand extends Entity {
+/**
+ * EntityArmorStand represents an armor stand entity.
+ *
+ * @author daoge_cmd
+ */
+public interface EntityArmorStand extends EntityLiving, EntityPhysicsComponent, EntityArmorStandBaseComponent, EntityContainerHolderComponent {
 
 }